SpreadsheetStorageContextTesting2 extracted SpreadsheetStorageContextTesting
